+++ backend/db/src/logging.rs
@@ -1,3 +1,5 @@
 mod pdf;
+mod web;
 
 pub use pdf::*;
+pub use web::*;