+++ backend/logging/src/lib.rs
@@ -1,5 +1,0 @@
-mod pdf;
-mod problem_set;
-mod web;
-
-pub use pdf::*;